feat(flask): create a new organization for project
-add estructure of path and files
