+++ src/__init__.py
@@ -1,0 +1,1 @@
+app = Flask(__name__)
