+++ src/routes/api_routes.py
@@ -1,0 +1,9 @@
+from src import app
+
+
+@app.route('/usuarios')
+def index():
+    cur = mysql.connection.cursor()
+    cur.execute("SELECT * FROM usuarios;")
+    data = cur.fetchall()
+    return render_template('index.html', usuarios=data)
